+++ tests/run-phan-tests.sh
@@ -38,11 +38,13 @@ sed -r -i "s/.\/tests\//.\/$TESTDIR\//" "$TESTDIR/bootstrap.php"
 sed -r -i "s/\/src\/Phan/\/vendor\/phan\/phan\/src\/Phan/" "$TESTDIR/bootstrap.php"
 sed -r -i "s/'\/src/'\/vendor\/phan\/phan\/src/" $TESTDIR/Phan/Language/UnionTypeTest.php
 
+# Enable verbose output, e.g. to know what tests were skipped
+sed -i 's/verbose="false"/verbose="true"/' phpunit.xml
 # Remove tests we don't care about
 EXCLUDE_PHPUNIT_TESTS="(ForkPoolTest|SoapTest)"
 sed -r -i "s/<file>.+$EXCLUDE_PHPUNIT_TESTS.*<\/file>//" phpunit.xml
 rm -r $TESTDIR/Phan/Language/Internal
-rm -rf $TESTDIR/Phan/{CLITest.php,PluginV3Test.php,Internal}
+rm -rf $TESTDIR/Phan/{CLITest.php,PluginV3Test.php,ForkPoolTest.php,SoapTest.php,Internal,LanguageServer}
 # Tests that rely on paths
 rm $TESTDIR/files/src/0545_require_testing.php $TESTDIR/Phan/Language/FileRefTest.php
 # Taint-check analyses the Group class earlier, thus avoiding a false positive issue from phan alone
@@ -57,6 +59,7 @@ composer dump-autoload
 
 SECCHECK_ISSUES="'SecurityCheck-DoubleEscaped', 'SecurityCheck-OTHER', 'SecurityCheck-SQLInjection', 'SecurityCheck-XSS', "
 SECCHECK_ISSUES+="'SecurityCheck-ShellInjection', 'SecurityCheck-PHPSerializeInjection', 'SecurityCheck-CUSTOM1', 'SecurityCheck-CUSTOM2', ";
+SECCHECK_ISSUES+="'SecurityCheck-PathTraversal', 'SecurityCheck-RCE', 'SecurityCheck-ReDoS', ";
 SECCHECK_ISSUES+="'SecurityCheckMulti', 'SecurityCheck-LikelyFalsePositive',";
 INNER_TEST_DIRS="misc/rewriting_test misc/fallback_test misc/config_override_test misc/empty_methods_plugin_test plugin_test real_types_test infer_missing_types_test"
 for DIR in $INNER_TEST_DIRS ; do
@@ -78,6 +81,26 @@ for DIR in $INNER_TEST_DIRS ; do
         sed -r -i "s/suppress_issue_types['\"] => \[/\0 $SECCHECK_ISSUES'/" $CFG_PATH
     else
         sed -r -i "s/^return \[/\0 'suppress_issue_types' => [ $SECCHECK_ISSUES ],/" $CFG_PATH
+    fi
+done
+
+# And also the generic config used for "base" tests. Note, some tests don't use a config file at all, so we can't
+# easily inject taint-check there.
+sed -r -i "s/^return \[/\0 'plugins' => [ '.\/MediaWikiSecurityCheckPlugin.php' ],/" $TESTDIR/.phan_for_test/config.php
+sed -r -i "s/'suppress_issue_types' => \[/\0 $SECCHECK_ISSUES/" $TESTDIR/.phan_for_test/config.php
+
+# Some are different...
+BASE_TESTS='PHP70Test PHP72Test PHP73Test PHP74Test PHP80Test'
+for BASE_TEST in $BASE_TESTS ; do
+    if grep -q "'plugins' =>" $TESTDIR/Phan/$BASE_TEST.php; then
+        sed -r -i "s/'plugins' => \[/\0'.\/MediaWikiSecurityCheckPlugin.php',/" $TESTDIR/Phan/$BASE_TEST.php
+    else
+        sed -r -i "s/OVERRIDES = \[/\0'plugins' => ['.\/MediaWikiSecurityCheckPlugin.php'],/" $TESTDIR/Phan/$BASE_TEST.php
+    fi
+    if grep -q "suppress_issue_types['\"] => \[" $TESTDIR/Phan/$BASE_TEST.php; then
+        sed -r -i "s/suppress_issue_types['\"] => \[/\0 $SECCHECK_ISSUES'/" $TESTDIR/Phan/$BASE_TEST.php
+    else
+        sed -r -i "s/OVERRIDES = \[/\0 'suppress_issue_types' => [ $SECCHECK_ISSUES ],/" $TESTDIR/Phan/$BASE_TEST.php
     fi
 done